+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.6] - 
+
+### Added
+
+- Allow filtering of metadata for profile
+
+## [0.1.5] - 2025-04-06
+
+### Added
+
+- Marketplace integration improvements
+
 ## [0.1.4] - 2025-04-04
 
 ### Fixed